Added simpleweight class and fixed return type

+++ src/main/java/weightScheme/weightScheme.java
@@ -5,5 +5,5 @@ import java.util.List;
 public interface weightScheme {
     public float computeWeight(List<Float> marks);
     public float getTotalWeight();
-    public float getNumberOfInstances();
+    public int getNumberOfInstances();
 }
